+++ CHANGES.md
@@ -1,5 +1,11 @@
 ## Release notes
 
+### Version 4.8.4 (20241025)
+* New version of Adminer [Releases notes](https://github.com/adminerevo/adminerevo/releases/tag/v4.8.4)
+* Add option to show a quick link in the navigation bar
+* Add small button to the modal header to open the content in a new tab without frame.
+* Changes the language of the Adminer ui suitable to the language of the current user.
+
 ### Version 4.8.3 (20231219)
 * New version of Adminer [Releases notes](https://github.com/adminerevo/adminerevo/releases/tag/v4.8.3)
 * MBS-8480: Fix modal width on error output (#16)